requirements.txt
fastapi[standard]
SQLAlchemy
passlib
python-jose
pydantic-settings
python-dotenv
psycopg2
bcrypt
python-multipart
boto3